fix(specs): fix tests

+++ src/app/components/shopping-cart-item/shopping-cart-item.component.ts
@@ -27,6 +27,7 @@ export class ShoppingCartItemComponent {
   }
 
   public removeFromCart(ticket: Ticket): void {
+    if (!this.cartItem?.event) return;
     this.shoppingCartService.removeFromCart({
       event: this.cartItem.event,
       ticket
